requirements.txt
langchain
streamlit
python-dotenv
pinecone-client
streamlit-chat
openai
tiktoken
langchain-openai
langchain-pinecone
